view model was implemented in settings feature module
# use cases were improved

+++ features/settings/src/main/java/com/vd/study/settings/domain/usecases/RemoveAccountUseCase.kt
@@ -1,8 +1,9 @@
 package com.vd.study.settings.domain.usecases
 
-import com.vd.study.core.dispatchers.IODispatcher
 import com.vd.study.core.container.Result
+import com.vd.study.core.dispatchers.IODispatcher
 import com.vd.study.settings.domain.entities.AccountEntity
+import com.vd.study.settings.domain.exceptions.IncorrectPasswordException
 import com.vd.study.settings.domain.repositories.SettingsRepository
 import kotlinx.coroutines.CoroutineDispatcher
 import kotlinx.coroutines.withContext
@@ -13,7 +14,10 @@ class RemoveAccountUseCase @Inject constructor(
     private val repository: SettingsRepository
 ) {
 
-    suspend operator fun invoke(account: AccountEntity): Result<Boolean> = withContext(ioDispatcher) {
-        return@withContext repository.removeAccount(account)
-    }
+    suspend operator fun invoke(account: AccountEntity, checkPassword: String): Result<Boolean> =
+        withContext(ioDispatcher) {
+            if (checkPassword != account.password) throw IncorrectPasswordException()
+
+            return@withContext repository.removeAccount(account)
+        }
 }
